Refactor unified turn loop into modular helpers
