+++ src/agent/runloop/unified/mod.rs
@@ -1,7 +1,15 @@
+mod curator;
+mod diagnostics;
 mod display;
+mod mcp_support;
+mod model_selection;
+mod palettes;
 mod prompts;
 mod session_setup;
 mod shell;
+mod status_line;
+mod tool_summary;
 mod turn;
+mod workspace_links;
 
 pub(crate) use turn::run_single_agent_loop_unified;
