+++ src/agent/runloop/unified/turn.rs
@@ -1,52 +1,30 @@
 use anyhow::{Context, Result, anyhow};
-use chrono::{DateTime, Local};
+use chrono::Local;
 use futures::StreamExt;
 use indicatif::ProgressStyle;
-use std::collections::{BTreeMap, BTreeSet, HashSet};
-use std::fmt::Write as FmtWrite;
-use std::fs;
-use std::io::{BufWriter, ErrorKind, Write};
+use std::collections::BTreeSet;
 use std::path::{Path, PathBuf};
-use std::process::Command;
-use std::str::FromStr;
 use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::time::{Duration, Instant};
-use tokio::io::{AsyncReadExt, AsyncWriteExt};
-use tokio::process::Command as TokioCommand;
 use tokio::sync::Notify;
 use tokio::sync::mpsc::UnboundedReceiver;
 use tokio::task;
 use tokio::time::sleep;
 
-use serde::Serialize;
-use serde_json::Value;
-use tempfile::Builder;
 use toml::Value as TomlValue;
 use tracing::warn;
 use vtcode_core::SimpleIndexer;
-use vtcode_core::config::api_keys::{ApiKeySources, get_api_key};
 use vtcode_core::config::constants::tools as tool_names;
 use vtcode_core::config::constants::{defaults, ui};
 use vtcode_core::config::loader::{ConfigManager, VTCodeConfig};
-use vtcode_core::config::mcp::McpTransportConfig;
-use vtcode_core::config::models::Provider;
-use vtcode_core::config::types::{AgentConfig as CoreAgentConfig, UiSurfacePreference};
+use vtcode_core::config::types::AgentConfig as CoreAgentConfig;
 use vtcode_core::core::agent::snapshots::{SnapshotConfig, SnapshotManager};
-use vtcode_core::config::{StatusLineConfig, StatusLineMode};
-use vtcode_core::core::context_curator::{
-    ConversationPhase, CuratedContext, Message as CuratorMessage,
-    ToolDefinition as CuratorToolDefinition,
-};
 use vtcode_core::core::decision_tracker::{Action as DTAction, DecisionOutcome, DecisionTracker};
 use vtcode_core::core::router::{Router, TaskClass};
 use vtcode_core::core::token_budget::{ContextComponent, TokenBudgetManager};
 use vtcode_core::llm::error_display;
-use vtcode_core::llm::factory::create_provider_with_config;
 use vtcode_core::llm::provider::{self as uni, LLMStreamEvent};
-use vtcode_core::llm::rig_adapter::{reasoning_parameters_for, verify_model_with_rig};
-use vtcode_core::mcp_client::McpClient;
-use vtcode_core::models::ModelId;
 use vtcode_core::tool_policy::ToolPolicy;
 use vtcode_core::tools::registry::{
     ToolErrorType, ToolExecutionError, ToolPermissionDecision, ToolRegistry,
@@ -54,12 +32,12 @@ use vtcode_core::tools::registry::{
 use vtcode_core::ui::slash::{SLASH_COMMANDS, SlashCommandInfo};
 use vtcode_core::ui::theme;
 use vtcode_core::ui::tui::{
-    InlineEvent, InlineHandle, InlineListItem, InlineListSelection, InlineTextStyle,
-    convert_style as convert_ui_style, spawn_session, theme_from_styles,
+    InlineEvent, InlineHandle, InlineTextStyle, convert_style as convert_ui_style, spawn_session,
+    theme_from_styles,
 };
 use vtcode_core::utils::ansi::{AnsiRenderer, MessageStyle};
 use vtcode_core::utils::session_archive::{
-    self, SessionArchive, SessionArchiveMetadata, SessionListing, SessionMessage,
+    self, SessionArchive, SessionArchiveMetadata, SessionMessage,
 };
 use vtcode_core::utils::transcript;
 
@@ -67,28 +45,41 @@ use crate::agent::runloop::ResumeSession;
 use crate::agent::runloop::context::{
     apply_aggressive_trim_unified, enforce_unified_context_window, prune_unified_tool_responses,
 };
-use crate::agent::runloop::git::{
-    GitStatusSummary, confirm_changes_with_git_diff, git_status_summary,
-};
+use crate::agent::runloop::git::confirm_changes_with_git_diff;
 use crate::agent::runloop::is_context_overflow_error;
-use crate::agent::runloop::model_picker::{
-    ModelPickerProgress, ModelPickerState, ModelSelectionResult,
-};
+use crate::agent::runloop::model_picker::{ModelPickerProgress, ModelPickerState};
 use crate::agent::runloop::prompt::refine_user_prompt_if_enabled;
 use crate::agent::runloop::slash_commands::{
-    McpCommandAction, SlashCommandOutcome, ThemePaletteMode, WorkspaceDirectoryCommand,
-    handle_slash_command,
+    McpCommandAction, SlashCommandOutcome, handle_slash_command,
 };
 use crate::agent::runloop::text_tools::{detect_textual_tool_call, extract_code_fence_blocks};
 use crate::agent::runloop::tool_output::render_code_fence_blocks;
 use crate::agent::runloop::tool_output::render_tool_output;
 use crate::agent::runloop::ui::{build_inline_header_context, render_session_banner};
 
+use super::curator::{
+    build_curated_sections, build_curator_messages, build_curator_tools, format_provider_label,
+    resolve_mode_label,
+};
+use super::diagnostics::run_doctor_diagnostics;
 use super::display::{display_user_message, ensure_turn_bottom_gap, persist_theme_preference};
+use super::mcp_support::{
+    display_mcp_providers, display_mcp_status, display_mcp_tools, refresh_mcp_tools,
+    render_mcp_login_guidance,
+};
+use super::model_selection::finalize_model_selection;
+use super::palettes::{
+    ActivePalette, apply_prompt_style, handle_palette_cancel, handle_palette_selection,
+    show_help_palette, show_sessions_palette, show_theme_palette,
+};
 use super::session_setup::{SessionState, initialize_session};
 use super::shell::{derive_recent_tool_output, should_short_circuit_shell};
+use super::status_line::{InputStatusState, update_input_status_if_changed};
+use super::tool_summary::{describe_tool_action, humanize_tool_name, render_tool_call_summary};
+use super::workspace_links::{
+    LinkedDirectory, handle_workspace_directory_command, remove_directory_symlink,
+};
 use crate::agent::runloop::mcp_events;
-use crate::agent::runloop::welcome::SessionBootstrap;
 
 #[derive(Default)]
 struct SessionStats {
@@ -103,217 +94,6 @@ impl SessionStats {
     fn sorted_tools(&self) -> Vec<String> {
         self.tools.iter().cloned().collect()
     }
-}
-
-#[derive(Clone)]
-struct LinkedDirectory {
-    original: PathBuf,
-    link_path: PathBuf,
-    display_path: String,
-}
-
-fn sanitize_alias_component(component: &str) -> String {
-    let lowered = component.trim().to_ascii_lowercase();
-    let mut sanitized = String::new();
-    let mut last_was_dash = false;
-
-    for ch in lowered.chars() {
-        if ch.is_ascii_alphanumeric() {
-            sanitized.push(ch);
-            last_was_dash = false;
-        } else if matches!(ch, '-' | '_' | '.' | ' ' | '/') {
-            if !last_was_dash && !sanitized.is_empty() {
-                sanitized.push('-');
-                last_was_dash = true;
-            }
-        }
-    }
-
-    let trimmed = sanitized.trim_matches('-').to_string();
-    if trimmed.is_empty() {
-        "linked-dir".to_string()
-    } else {
-        trimmed
-    }
-}
-
-fn create_directory_symlink(target: &Path, link: &Path) -> Result<()> {
-    #[cfg(unix)]
-    {
-        std::os::unix::fs::symlink(target, link).with_context(|| {
-            format!(
-                "failed to create symlink {} -> {}",
-                link.display(),
-                target.display()
-            )
-        })?;
-    }
-
-    #[cfg(windows)]
-    {
-        std::os::windows::fs::symlink_dir(target, link).with_context(|| {
-            format!(
-                "failed to create symlink {} -> {}",
-                link.display(),
-                target.display()
-            )
-        })?;
-    }
-
-    Ok(())
-}
-
-fn remove_directory_symlink(path: &Path) -> Result<()> {
-    #[cfg(unix)]
-    {
-        if let Err(err) = fs::remove_file(path) {
-            if err.kind() != ErrorKind::NotFound {
-                return Err(err).with_context(|| {
-                    format!("failed to remove directory link {}", path.display())
-                });
-            }
-        }
-    }
-
-    #[cfg(windows)]
-    {
-        if let Err(err) = fs::remove_dir(path) {
-            if err.kind() != ErrorKind::NotFound {
-                return Err(err).with_context(|| {
-                    format!("failed to remove directory link {}", path.display())
-                });
-            }
-        }
-    }
-
-    Ok(())
-}
-
-const THEME_PALETTE_TITLE: &str = "Theme picker";
-const THEME_LIST_TITLE: &str = "Available themes";
-const THEME_ACTIVE_BADGE: &str = "Active";
-const THEME_SELECT_HINT: &str = "Use ↑/↓ to choose a theme, Enter to apply, Esc to cancel.";
-const THEME_INSPECT_HINT: &str = "Use ↑/↓ to browse themes, Enter to view details, Esc to close.";
-const SESSIONS_PALETTE_TITLE: &str = "Archived sessions";
-const SESSIONS_HINT_PRIMARY: &str = "Use ↑/↓ to browse sessions.";
-const SESSIONS_HINT_SECONDARY: &str = "Enter to print details • Esc to close.";
-const SESSIONS_LATEST_BADGE: &str = "Latest";
-const HELP_PALETTE_TITLE: &str = "Slash command help";
-const HELP_HINT_PRIMARY: &str = "Use ↑/↓ to pick a slash command.";
-const HELP_HINT_SECONDARY: &str = "Enter to insert into the input • Esc to dismiss.";
-
-enum ActivePalette {
-    Theme {
-        mode: ThemePaletteMode,
-    },
-    Sessions {
-        listings: Vec<SessionListing>,
-        limit: usize,
-    },
-    Help,
-}
-
-fn show_theme_palette(renderer: &mut AnsiRenderer, mode: ThemePaletteMode) -> Result<bool> {
-    let title = match mode {
-        ThemePaletteMode::Select => THEME_PALETTE_TITLE,
-        ThemePaletteMode::Inspect => THEME_LIST_TITLE,
-    };
-    let hint = match mode {
-        ThemePaletteMode::Select => THEME_SELECT_HINT,
-        ThemePaletteMode::Inspect => THEME_INSPECT_HINT,
-    };
-
-    let current_id = theme::active_theme_id();
-    let current_label = theme::active_theme_label().to_string();
-    let mut items = Vec::new();
-
-    for id in theme::available_themes() {
-        let label = theme::theme_label(id).unwrap_or(id);
-        let badge = (id == current_id).then(|| THEME_ACTIVE_BADGE.to_string());
-        items.push(InlineListItem {
-            title: label.to_string(),
-            subtitle: Some(format!("ID: {}", id)),
-            badge,
-            indent: 0,
-            selection: Some(InlineListSelection::Theme(id.to_string())),
-            search_value: None,
-        });
-    }
-
-    if items.is_empty() {
-        renderer.line(MessageStyle::Info, "No themes available.")?;
-        return Ok(false);
-    }
-
-    let lines = vec![
-        format!("Current theme: {}", current_label),
-        hint.to_string(),
-    ];
-    renderer.show_list_modal(
-        title,
-        lines,
-        items,
-        Some(InlineListSelection::Theme(current_id)),
-        None,
-    );
-
-    Ok(true)
-}
-
-fn show_sessions_palette(
-    renderer: &mut AnsiRenderer,
-    listings: &[SessionListing],
-    limit: usize,
-) -> Result<bool> {
-    if listings.is_empty() {
-        renderer.line(MessageStyle::Info, "No archived sessions found.")?;
-        return Ok(false);
-    }
-
-    let mut items = Vec::new();
-    for (index, listing) in listings.iter().enumerate() {
-        let ended_local = listing
-            .snapshot
-            .ended_at
-            .with_timezone(&Local)
-            .format("%Y-%m-%d %H:%M");
-        let duration = listing
-            .snapshot
-            .ended_at
-            .signed_duration_since(listing.snapshot.started_at);
-        let duration_std = duration.to_std().unwrap_or_else(|_| Duration::from_secs(0));
-        let detail = format!(
-            "Duration: {} · {} msgs · {} tools",
-            format_duration_label(duration_std),
-            listing.snapshot.total_messages,
-            listing.snapshot.distinct_tools.len(),
-        );
-        let badge = (index == 0).then(|| SESSIONS_LATEST_BADGE.to_string());
-        items.push(InlineListItem {
-            title: format!(
-                "{} · {} · {}",
-                ended_local,
-                listing.snapshot.metadata.model,
-                listing.snapshot.metadata.workspace_label,
-            ),
-            subtitle: Some(detail),
-            badge,
-            indent: 0,
-            selection: Some(InlineListSelection::Session(listing.identifier())),
-            search_value: None,
-        });
-    }
-
-    let lines = vec![
-        format!("Showing {} of {} archived sessions", listings.len(), limit),
-        SESSIONS_HINT_PRIMARY.to_string(),
-        SESSIONS_HINT_SECONDARY.to_string(),
-    ];
-    let selected = listings
-        .first()
-        .map(|listing| InlineListSelection::Session(listing.identifier()));
-    renderer.show_list_modal(SESSIONS_PALETTE_TITLE, lines, items, selected, None);
-    Ok(true)
 }
 
 async fn display_session_status(
@@ -417,1418 +197,6 @@ async fn display_token_cost(
     Ok(())
 }
 
-fn show_help_palette(
-    renderer: &mut AnsiRenderer,
-    commands: &[&'static SlashCommandInfo],
-) -> Result<bool> {
-    if commands.is_empty() {
-        renderer.line(MessageStyle::Info, "No slash commands available.")?;
-        return Ok(false);
-    }
-
-    let mut items = Vec::new();
-    for info in commands {
-        let subtitle = if info.description.is_empty() {
-            None
-        } else {
-            Some(info.description.to_string())
-        };
-        items.push(InlineListItem {
-            title: format!("/{}", info.name),
-            subtitle,
-            badge: None,
-            indent: 0,
-            selection: Some(InlineListSelection::SlashCommand(info.name.to_string())),
-            search_value: None,
-        });
-    }
-
-    let lines = vec![
-        HELP_HINT_PRIMARY.to_string(),
-        HELP_HINT_SECONDARY.to_string(),
-    ];
-    let selected = commands
-        .first()
-        .map(|info| InlineListSelection::SlashCommand(info.name.to_string()));
-    renderer.show_list_modal(HELP_PALETTE_TITLE, lines, items, selected, None);
-    Ok(true)
-}
-
-fn render_session_details(renderer: &mut AnsiRenderer, listing: &SessionListing) -> Result<()> {
-    let ended_local = listing
-        .snapshot
-        .ended_at
-        .with_timezone(&Local)
-        .format("%Y-%m-%d %H:%M");
-    let duration = listing
-        .snapshot
-        .ended_at
-        .signed_duration_since(listing.snapshot.started_at);
-    let duration_std = duration.to_std().unwrap_or_else(|_| Duration::from_secs(0));
-    let duration_label = format_duration_label(duration_std);
-    let tool_count = listing.snapshot.distinct_tools.len();
-
-    renderer.line(
-        MessageStyle::Info,
-        &format!(
-            "- (ID: {}) {} · Model: {} · Workspace: {}",
-            listing.identifier(),
-            ended_local,
-            listing.snapshot.metadata.model,
-            listing.snapshot.metadata.workspace_label,
-        ),
-    )?;
-    renderer.line(
-        MessageStyle::Info,
-        &format!(
-            "    Duration: {} · {} msgs · {} tools",
-            duration_label, listing.snapshot.total_messages, tool_count,
-        ),
-    )?;
-
-    if let Some(prompt) = listing.first_prompt_preview() {
-        renderer.line(MessageStyle::Info, &format!("    Prompt: {prompt}"))?;
-    }
-
-    if let Some(reply) = listing.first_reply_preview() {
-        renderer.line(MessageStyle::Info, &format!("    Reply: {reply}"))?;
-    }
-
-    renderer.line(
-        MessageStyle::Info,
-        &format!("    File: {}", listing.path.display()),
-    )?;
-    Ok(())
-}
-
-fn format_duration_label(duration: Duration) -> String {
-    let total_seconds = duration.as_secs();
-    let hours = total_seconds / 3600;
-    let minutes = (total_seconds % 3600) / 60;
-    let seconds = total_seconds % 60;
-
-    let mut parts = Vec::new();
-    if hours > 0 {
-        parts.push(format!("{}h", hours));
-    }
-    if minutes > 0 || hours > 0 {
-        parts.push(format!("{}m", minutes));
-    }
-    parts.push(format!("{}s", seconds));
-    parts.join(" ")
-}
-
-fn handle_workspace_directory_command(
-    renderer: &mut AnsiRenderer,
-    workspace_root: &Path,
-    command: WorkspaceDirectoryCommand,
-    linked_directories: &mut Vec<LinkedDirectory>,
-) -> Result<()> {
-    match command {
-        WorkspaceDirectoryCommand::List => {
-            renderer.line(MessageStyle::Status, "Linked directories:")?;
-            if linked_directories.is_empty() {
-                renderer.line(MessageStyle::Info, "  (none)")?;
-            } else {
-                for (index, entry) in linked_directories.iter().enumerate() {
-                    renderer.line(
-                        MessageStyle::Info,
-                        &format!(
-                            "  {}. {} → {}",
-                            index + 1,
-                            entry.display_path,
-                            entry.original.display()
-                        ),
-                    )?;
-                }
-            }
-            renderer.line(
-                MessageStyle::Info,
-                "Use /add-dir <path> to add or /add-dir --remove <alias> to detach.",
-            )?;
-            Ok(())
-        }
-        WorkspaceDirectoryCommand::Add(raw_paths) => {
-            if raw_paths.is_empty() {
-                return Ok(());
-            }
-
-            let link_root = workspace_root.join(".vtcode").join("external");
-            fs::create_dir_all(&link_root).with_context(|| {
-                format!("failed to prepare link directory {}", link_root.display())
-            })?;
-
-            for raw in raw_paths {
-                let trimmed = raw.trim();
-                if trimmed.is_empty() {
-                    continue;
-                }
-
-                let candidate = PathBuf::from(trimmed);
-                let resolved = if candidate.is_absolute() {
-                    candidate
-                } else {
-                    workspace_root.join(candidate)
-                };
-
-                let canonical = match fs::canonicalize(&resolved) {
-                    Ok(path) => path,
-                    Err(err) => {
-                        renderer.line(
-                            MessageStyle::Error,
-                            &format!("Failed to resolve '{}': {}", resolved.display(), err),
-                        )?;
-                        continue;
-                    }
-                };
-
-                if !canonical.is_dir() {
-                    renderer.line(
-                        MessageStyle::Error,
-                        &format!(
-                            "Path '{}' is not a directory and cannot be linked.",
-                            canonical.display()
-                        ),
-                    )?;
-                    continue;
-                }
-
-                if linked_directories
-                    .iter()
-                    .any(|entry| entry.original == canonical)
-                {
-                    renderer.line(
-                        MessageStyle::Info,
-                        &format!("Directory already linked: {}", canonical.display()),
-                    )?;
-                    continue;
-                }
-
-                let alias_base = canonical
-                    .file_name()
-                    .and_then(|value| value.to_str())
-                    .map(sanitize_alias_component)
-                    .filter(|alias| !alias.is_empty())
-                    .unwrap_or_else(|| format!("linked-dir-{}", linked_directories.len() + 1));
-
-                let mut alias = alias_base.clone();
-                let mut counter = 2usize;
-                while linked_directories
-                    .iter()
-                    .any(|entry| entry.display_path.ends_with(&alias))
-                    || link_root.join(&alias).exists()
-                {
-                    alias = format!("{}-{}", alias_base, counter);
-                    counter += 1;
-                }
-
-                let link_path = link_root.join(&alias);
-                if let Err(err) = create_directory_symlink(&canonical, &link_path) {
-                    renderer.line(
-                        MessageStyle::Error,
-                        &format!("Failed to link {}: {}", canonical.display(), err),
-                    )?;
-                    continue;
-                }
-
-                let display_path = format!(".vtcode/external/{}", alias);
-                renderer.line(
-                    MessageStyle::Info,
-                    &format!("Linked {} as {}", canonical.display(), display_path),
-                )?;
-                renderer.line(
-                    MessageStyle::Info,
-                    "Access files in this directory using the linked path inside the workspace.",
-                )?;
-
-                linked_directories.push(LinkedDirectory {
-                    original: canonical,
-                    link_path,
-                    display_path,
-                });
-            }
-
-            Ok(())
-        }
-        WorkspaceDirectoryCommand::Remove(targets) => {
-            if targets.is_empty() {
-                renderer.line(
-                    MessageStyle::Error,
-                    "Usage: /add-dir --remove <alias|path> [more...]",
-                )?;
-                return Ok(());
-            }
-
-            let mut any_removed = false;
-            for target in targets {
-                match detach_linked_directory(renderer, &target, linked_directories) {
-                    Ok(true) => {
-                        any_removed = true;
-                    }
-                    Ok(false) => {}
-                    Err(err) => {
-                        renderer.line(
-                            MessageStyle::Error,
-                            &format!("Failed to remove '{}': {}", target, err),
-                        )?;
-                    }
-                }
-            }
-
-            if !any_removed {
-                renderer.line(
-                    MessageStyle::Info,
-                    "No matching linked directories were removed.",
-                )?;
-            }
-
-            Ok(())
-        }
-    }
-}
-
-fn detach_linked_directory(
-    renderer: &mut AnsiRenderer,
-    target: &str,
-    linked_directories: &mut Vec<LinkedDirectory>,
-) -> Result<bool> {
-    if linked_directories.is_empty() {
-        renderer.line(MessageStyle::Info, "No linked directories to remove.")?;
-        return Ok(false);
-    }
-
-    let normalized = target.trim();
-    if normalized.is_empty() {
-        return Ok(false);
-    }
-
-    let stripped = normalized
-        .strip_prefix(".vtcode/external/")
-        .unwrap_or(normalized)
-        .to_string();
-
-    let mut index_to_remove: Option<usize> = None;
-
-    if let Ok(number) = stripped.parse::<usize>() {
-        if number >= 1 && number <= linked_directories.len() {
-            index_to_remove = Some(number - 1);
-        }
-    }
-
-    if index_to_remove.is_none() {
-        for (index, entry) in linked_directories.iter().enumerate() {
-            if entry.display_path.ends_with(&stripped)
-                || entry
-                    .original
-                    .to_string_lossy()
-                    .eq_ignore_ascii_case(normalized)
-                || entry
-                    .link_path
-                    .to_string_lossy()
-                    .eq_ignore_ascii_case(normalized)
-            {
-                index_to_remove = Some(index);
-                break;
-            }
-        }
-    }
-
-    if let Some(index) = index_to_remove {
-        let entry = linked_directories.remove(index);
-        remove_directory_symlink(&entry.link_path)?;
-        renderer.line(
-            MessageStyle::Info,
-            &format!("Removed linked directory {}", entry.display_path),
-        )?;
-        return Ok(true);
-    }
-
-    renderer.line(
-        MessageStyle::Error,
-        &format!("No linked directory matched '{}'.", normalized),
-    )?;
-    Ok(false)
-}
-
-async fn display_mcp_status(
-    renderer: &mut AnsiRenderer,
-    session_bootstrap: &SessionBootstrap,
-    tool_registry: &mut ToolRegistry,
-    mcp_client: Option<&Arc<McpClient>>,
-    mcp_panel_state: &mcp_events::McpPanelState,
-) -> Result<()> {
-    renderer.line(MessageStyle::Status, "MCP status overview:")?;
-
-    match session_bootstrap.mcp_enabled {
-        Some(true) => renderer.line(MessageStyle::Status, "  • Enabled in vtcode.toml")?,
-        Some(false) => renderer.line(MessageStyle::Status, "  • Disabled in vtcode.toml")?,
-        None => renderer.line(MessageStyle::Status, "  • No MCP configuration detected")?,
-    }
-
-    if let Some(providers) = &session_bootstrap.mcp_providers {
-        if providers.is_empty() {
-            renderer.line(MessageStyle::Status, "  • No MCP providers configured")?;
-        } else {
-            let enabled: Vec<String> = providers
-                .iter()
-                .filter(|provider| provider.enabled)
-                .map(|provider| provider.name.clone())
-                .collect();
-            if enabled.is_empty() {
-                renderer.line(
-                    MessageStyle::Status,
-                    "  • Providers present but all disabled",
-                )?;
-            } else {
-                renderer.line(
-                    MessageStyle::Status,
-                    &format!("  • Enabled providers: {}", enabled.join(", ")),
-                )?;
-            }
-        }
-    }
-
-    if let Some(client) = mcp_client {
-        let status = client.get_status();
-        renderer.line(
-            MessageStyle::Status,
-            &format!(
-                "  • Runtime connections: {} active / {} configured",
-                status.active_connections, status.provider_count
-            ),
-        )?;
-
-        if !status.configured_providers.is_empty() {
-            renderer.line(
-                MessageStyle::Status,
-                &format!(
-                    "  • Configured providers: {}",
-                    status.configured_providers.join(", ")
-                ),
-            )?;
-        }
-
-        match tool_registry.list_mcp_tools().await {
-            Ok(tools) => {
-                if tools.is_empty() {
-                    renderer.line(
-                        MessageStyle::Status,
-                        "  • No MCP tools exposed by providers",
-                    )?;
-                } else {
-                    let mut samples = Vec::new();
-                    for info in tools.iter().take(5) {
-                        samples.push(format!("{} ({})", info.name, info.provider));
-                    }
-                    let extra = tools.len().saturating_sub(samples.len());
-                    let suffix = if extra > 0 {
-                        format!(" and {} more", extra)
-                    } else {
-                        String::new()
-                    };
-                    renderer.line(
-                        MessageStyle::Status,
-                        &format!("  • Tools available: {}{}", samples.join(", "), suffix),
-                    )?;
-                }
-            }
-            Err(err) => {
-                renderer.line(
-                    MessageStyle::Error,
-                    &format!("  • Failed to list MCP tools: {}", err),
-                )?;
-            }
-        }
-    } else {
-        renderer.line(
-            MessageStyle::Status,
-            "  • MCP client inactive in this session",
-        )?;
-    }
-
-    if mcp_panel_state.is_enabled() {
-        let recent = mcp_panel_state.recent_events_snapshot(5);
-        if !recent.is_empty() {
-            renderer.line(MessageStyle::McpStatus, "Recent MCP activity:")?;
-            for event in recent {
-                let timestamp: DateTime<Local> = DateTime::<Local>::from(event.timestamp);
-                let mut detail = event.status.label().to_string();
-                if let Some(args) = event.args_preview.as_ref() {
-                    if !args.trim().is_empty() {
-                        let preview: String = args.chars().take(80).collect();
-                        if preview.len() < args.len() {
-                            detail.push_str(&format!(" · args {}…", preview));
-                        } else {
-                            detail.push_str(&format!(" · args {}", preview));
-                        }
-                    }
-                }
-
-                renderer.line(
-                    MessageStyle::McpStatus,
-                    &format!(
-                        "    {} [{}] {}::{} — {}",
-                        event.status.symbol(),
-                        timestamp.format("%H:%M:%S"),
-                        event.provider,
-                        event.method,
-                        detail
-                    ),
-                )?;
-            }
-        }
-    }
-
-    renderer.line(
-        MessageStyle::Info,
-        "Use `vtcode mcp list` to review and manage providers.",
-    )?;
-    Ok(())
-}
-
-fn display_mcp_providers(
-    renderer: &mut AnsiRenderer,
-    session_bootstrap: &SessionBootstrap,
-    mcp_client: Option<&Arc<McpClient>>,
-) -> Result<()> {
-    renderer.line(
-        MessageStyle::Status,
-        "MCP providers configured in vtcode.toml:",
-    )?;
-
-    let Some(configured) = &session_bootstrap.mcp_providers else {
-        renderer.line(
-            MessageStyle::Info,
-            "No vtcode.toml configuration detected for MCP providers.",
-        )?;
-        renderer.line(
-            MessageStyle::Info,
-            "Use `vtcode mcp add <name> --command <...>` to register providers.",
-        )?;
-        return Ok(());
-    };
-
-    if configured.is_empty() {
-        renderer.line(MessageStyle::Info, "No providers defined in vtcode.toml.")?;
-        renderer.line(
-            MessageStyle::Info,
-            "Run `vtcode mcp add` to connect Claude Code to external tools.",
-        )?;
-        return Ok(());
-    }
-
-    let active = mcp_client
-        .map(|client| {
-            client
-                .get_status()
-                .configured_providers
-                .into_iter()
-                .collect::<HashSet<_>>()
-        })
-        .unwrap_or_default();
-
-    for provider in configured {
-        let runtime_state = if active.contains(&provider.name) {
-            "connected"
-        } else if provider.enabled {
-            "inactive"
-        } else {
-            "disabled"
-        };
-        let status_label = if provider.enabled {
-            "enabled"
-        } else {
-            "disabled"
-        };
-        renderer.line(
-            MessageStyle::Info,
-            &format!("- {} ({}, {})", provider.name, status_label, runtime_state),
-        )?;
-
-        match &provider.transport {
-            McpTransportConfig::Stdio(stdio) => {
-                let mut command_desc = stdio.command.clone();
-                if !stdio.args.is_empty() {
-                    command_desc.push(' ');
-                    command_desc.push_str(&stdio.args.join(" "));
-                }
-                renderer.line(
-                    MessageStyle::Info,
-                    &format!("    transport: stdio · {}", command_desc),
-                )?;
-                if let Some(dir) = &stdio.working_directory {
-                    renderer.line(MessageStyle::Info, &format!("    working_dir: {}", dir))?;
-                }
-            }
-            McpTransportConfig::Http(http) => {
-                renderer.line(
-                    MessageStyle::Info,
-                    &format!("    transport: http · {}", http.endpoint),
-                )?;
-                if let Some(env) = &http.api_key_env {
-                    renderer.line(
-                        MessageStyle::Info,
-                        &format!("    bearer token env: {}", env),
-                    )?;
-                }
-                if !http.headers.is_empty() {
-                    renderer.line(
-                        MessageStyle::Info,
-                        &format!(
-                            "    headers: {}",
-                            http.headers.keys().cloned().collect::<Vec<_>>().join(", ")
-                        ),
-                    )?;
-                }
-            }
-        }
-
-        renderer.line(
-            MessageStyle::Info,
-            &format!(
-                "    max concurrent requests: {}",
-                provider.max_concurrent_requests
-            ),
-        )?;
-
-        if !provider.env.is_empty() {
-            renderer.line(
-                MessageStyle::Info,
-                &format!(
-                    "    env vars: {}",
-                    provider.env.keys().cloned().collect::<Vec<_>>().join(", ")
-                ),
-            )?;
-        }
-    }
-
-    renderer.line(
-        MessageStyle::Info,
-        "Use `vtcode mcp get <name>` for full JSON output or `vtcode mcp remove <name>` to delete.",
-    )?;
-    Ok(())
-}
-
-async fn display_mcp_tools(
-    renderer: &mut AnsiRenderer,
-    tool_registry: &mut ToolRegistry,
-) -> Result<()> {
-    renderer.line(
-        MessageStyle::Status,
-        "Listing MCP tools exposed by connected providers:",
-    )?;
-    match tool_registry.list_mcp_tools().await {
-        Ok(tools) => {
-            if tools.is_empty() {
-                renderer.line(
-                    MessageStyle::Info,
-                    "No MCP tools are currently available. Try /mcp refresh after connecting providers.",
-                )?;
-                return Ok(());
-            }
-
-            let mut grouped: BTreeMap<String, Vec<_>> = BTreeMap::new();
-            for tool in tools {
-                grouped.entry(tool.provider.clone()).or_default().push(tool);
-            }
-
-            for (provider, mut entries) in grouped {
-                entries.sort_by(|a, b| a.name.cmp(&b.name));
-                renderer.line(
-                    MessageStyle::Info,
-                    &format!("- Provider: {} ({} tool(s))", provider, entries.len()),
-                )?;
-                for info in entries {
-                    renderer.line(MessageStyle::Info, &format!("    • {}", info.name))?;
-                }
-            }
-        }
-        Err(err) => {
-            renderer.line(
-                MessageStyle::Error,
-                &format!("Failed to list MCP tools: {}", err),
-            )?;
-        }
-    }
-    Ok(())
-}
-
-async fn refresh_mcp_tools(
-    renderer: &mut AnsiRenderer,
-    tool_registry: &mut ToolRegistry,
-) -> Result<()> {
-    renderer.line(MessageStyle::Status, "Refreshing MCP tool index…")?;
-    match tool_registry.refresh_mcp_tools().await {
-        Ok(()) => {
-            match tool_registry.list_mcp_tools().await {
-                Ok(tools) => {
-                    renderer.line(
-                        MessageStyle::Info,
-                        &format!("Indexed {} MCP tool(s).", tools.len()),
-                    )?;
-                }
-                Err(err) => {
-                    renderer.line(
-                        MessageStyle::Error,
-                        &format!("Refreshed but failed to list tools: {}", err),
-                    )?;
-                }
-            }
-            renderer.line(
-                MessageStyle::Info,
-                "Use /mcp tools to inspect the refreshed catalog.",
-            )?;
-        }
-        Err(err) => {
-            renderer.line(
-                MessageStyle::Error,
-                &format!("Failed to refresh MCP tools: {}", err),
-            )?;
-        }
-    }
-    Ok(())
-}
-
-fn render_mcp_login_guidance(
-    renderer: &mut AnsiRenderer,
-    provider: String,
-    is_login: bool,
-) -> Result<()> {
-    let trimmed = provider.trim();
-    if trimmed.is_empty() {
-        renderer.line(
-            MessageStyle::Error,
-            "Provider name required. Usage: /mcp login <name> or /mcp logout <name>.",
-        )?;
-        return Ok(());
-    }
-
-    let action = if is_login { "login" } else { "logout" };
-    renderer.line(
-        MessageStyle::Status,
-        &format!(
-            "OAuth {} flow requested for provider '{}'.",
-            action, trimmed
-        ),
-    )?;
-    renderer.line(
-        MessageStyle::Info,
-        "VTCode delegates OAuth to the CLI today.",
-    )?;
-    renderer.line(
-        MessageStyle::Info,
-        &format!("Run: vtcode mcp {} {}", action, trimmed),
-    )?;
-    renderer.line(
-        MessageStyle::Info,
-        "This command will walk you through the authentication flow in your shell.",
-    )?;
-    Ok(())
-}
-
-fn detect_command_version(command: &str, args: &[&str]) -> std::result::Result<String, String> {
-    match Command::new(command).args(args).output() {
-        Ok(output) => {
-            if output.status.success() {
-                let stdout = String::from_utf8_lossy(&output.stdout).trim().to_string();
-                let stderr = String::from_utf8_lossy(&output.stderr).trim().to_string();
-                let version_text = if !stdout.is_empty() { stdout } else { stderr };
-                if version_text.is_empty() {
-                    Err(format!(
-                        "{} {} returned no version output",
-                        command,
-                        args.join(" ")
-                    ))
-                } else {
-                    Ok(version_text)
-                }
-            } else {
-                let stderr = String::from_utf8_lossy(&output.stderr).trim().to_string();
-                let reason = if !stderr.is_empty() {
-                    stderr
-                } else {
-                    output.status.to_string()
-                };
-                Err(format!(
-                    "{} {} exited with error: {}",
-                    command,
-                    args.join(" "),
-                    reason
-                ))
-            }
-        }
-        Err(err) => {
-            if err.kind() == ErrorKind::NotFound {
-                Err(format!("{} not found in PATH", command))
-            } else {
-                Err(format!("Failed to execute {}: {}", command, err))
-            }
-        }
-    }
-}
-
-fn run_doctor_diagnostics(
-    renderer: &mut AnsiRenderer,
-    config: &CoreAgentConfig,
-    vt_cfg: Option<&VTCodeConfig>,
-    provider_name: &str,
-    mcp_client: Option<&Arc<McpClient>>,
-    linked_directories: &[LinkedDirectory],
-) -> Result<()> {
-    renderer.line(MessageStyle::Status, "Running VTCode doctor checks:")?;
-
-    let workspace_result = if config.workspace.is_dir() {
-        Ok(format!("{}", config.workspace.display()))
-    } else {
-        Err("Workspace directory is missing or inaccessible".to_string())
-    };
-    render_doctor_check(renderer, "Workspace", workspace_result)?;
-
-    let config_result = match ConfigManager::load_from_workspace(&config.workspace) {
-        Ok(manager) => {
-            if let Some(path) = manager.config_path() {
-                Ok(format!("Loaded configuration from {}", path.display()))
-            } else {
-                Ok("Using runtime defaults (no vtcode.toml found)".to_string())
-            }
-        }
-        Err(err) => Err(err.to_string()),
-    };
-    render_doctor_check(renderer, "Configuration", config_result)?;
-
-    let provider_label = if provider_name.trim().is_empty() {
-        "gemini"
-    } else {
-        provider_name
-    };
-    let api_key_result = match get_api_key(provider_label, &ApiKeySources::default()) {
-        Ok(_) => Ok(format!(
-            "API key detected for provider '{}'.",
-            provider_label
-        )),
-        Err(err) => Err(format!(
-            "Missing API key for provider '{}': {}",
-            provider_label, err
-        )),
-    };
-    render_doctor_check(renderer, "API key", api_key_result)?;
-
-    let cli_version = format!("VTCode {}", env!("CARGO_PKG_VERSION"));
-    render_doctor_check(renderer, "VTCode CLI", Ok(cli_version))?;
-
-    let node_result = detect_command_version("node", &["--version"])
-        .map(|version| format!("Found Node.js {}", version));
-    render_doctor_check(renderer, "Node.js", node_result)?;
-
-    let npm_result = detect_command_version("npm", &["--version"])
-        .map(|version| format!("Found npm {}", version));
-    render_doctor_check(renderer, "npm", npm_result)?;
-
-    let claude_result = detect_command_version("claude", &["--version"])
-        .map(|version| format!("Found Claude CLI {}", version));
-    render_doctor_check(renderer, "Claude CLI", claude_result)?;
-
-    let mcp_result = if let Some(cfg) = vt_cfg {
-        if cfg.mcp.enabled {
-            if let Some(client) = mcp_client {
-                let status = client.get_status();
-                Ok(format!(
-                    "Enabled with {} configured provider(s), {} active",
-                    status.configured_providers.len(),
-                    status.active_connections
-                ))
-            } else {
-                Err("Enabled in configuration but client failed to initialize".to_string())
-            }
-        } else {
-            Ok("Disabled in configuration".to_string())
-        }
-    } else {
-        Ok("No MCP configuration detected".to_string())
-    };
-    render_doctor_check(renderer, "MCP", mcp_result)?;
-
-    if linked_directories.is_empty() {
-        renderer.line(MessageStyle::Status, "Linked directories: none")?;
-    } else {
-        let aliases: Vec<String> = linked_directories
-            .iter()
-            .map(|entry| entry.display_path.clone())
-            .collect();
-        renderer.line(
-            MessageStyle::Status,
-            &format!("Linked directories: {}", aliases.join(", ")),
-        )?;
-    }
-
-    renderer.line(
-        MessageStyle::Info,
-        "Doctor finished. Run `cargo check` or `vtcode mcp list` for more details if needed.",
-    )?;
-    Ok(())
-}
-
-fn render_doctor_check(
-    renderer: &mut AnsiRenderer,
-    label: &str,
-    outcome: std::result::Result<String, String>,
-) -> Result<()> {
-    match outcome {
-        Ok(detail) => {
-            renderer.line(MessageStyle::Status, &format!("[ok] {}: {}", label, detail))?
-        }
-        Err(detail) => renderer.line(
-            MessageStyle::Error,
-            &format!("[fail] {}: {}", label, detail),
-        )?,
-    }
-    Ok(())
-}
-
-fn handle_palette_selection(
-    palette: ActivePalette,
-    selection: InlineListSelection,
-    renderer: &mut AnsiRenderer,
-    handle: &InlineHandle,
-) -> Result<Option<ActivePalette>> {
-    match palette {
-        ActivePalette::Theme { mode } => match selection {
-            InlineListSelection::Theme(theme_id) => match mode {
-                ThemePaletteMode::Select => {
-                    match theme::set_active_theme(&theme_id) {
-                        Ok(()) => {
-                            let label = theme::active_theme_label();
-                            renderer.line(
-                                MessageStyle::Info,
-                                &format!("Theme switched to {}", label),
-                            )?;
-                            persist_theme_preference(renderer, &theme_id)?;
-                            let styles = theme::active_styles();
-                            handle.set_theme(theme_from_styles(&styles));
-                            apply_prompt_style(handle);
-                        }
-                        Err(err) => {
-                            renderer.line(
-                                MessageStyle::Error,
-                                &format!("Theme '{}' not available: {}", theme_id, err),
-                            )?;
-                        }
-                    }
-                    Ok(None)
-                }
-                ThemePaletteMode::Inspect => {
-                    let label = theme::theme_label(&theme_id).unwrap_or_else(|| theme_id.as_str());
-                    renderer.line(
-                        MessageStyle::Info,
-                        &format!("Theme {} ({}) is available.", label, theme_id),
-                    )?;
-                    if show_theme_palette(renderer, mode)? {
-                        Ok(Some(ActivePalette::Theme { mode }))
-                    } else {
-                        Ok(None)
-                    }
-                }
-            },
-            _ => Ok(Some(ActivePalette::Theme { mode })),
-        },
-        ActivePalette::Sessions { listings, limit } => {
-            if let InlineListSelection::Session(selected_id) = &selection {
-                if let Some(listing) = listings
-                    .iter()
-                    .find(|entry| entry.identifier() == *selected_id)
-                    .cloned()
-                {
-                    render_session_details(renderer, &listing)?;
-                }
-            }
-            if show_sessions_palette(renderer, &listings, limit)? {
-                Ok(Some(ActivePalette::Sessions { listings, limit }))
-            } else {
-                Ok(None)
-            }
-        }
-        ActivePalette::Help => {
-            if let InlineListSelection::SlashCommand(command) = selection {
-                handle.set_input(format!("/{} ", command));
-                renderer.line(
-                    MessageStyle::Info,
-                    &format!("Inserted '/{}' into the input.", command),
-                )?;
-            }
-            Ok(None)
-        }
-    }
-}
-
-fn handle_palette_cancel(palette: ActivePalette, renderer: &mut AnsiRenderer) -> Result<()> {
-    match palette {
-        ActivePalette::Theme { mode } => {
-            let message = match mode {
-                ThemePaletteMode::Select => "Theme selection cancelled.",
-                ThemePaletteMode::Inspect => "Closed theme list.",
-            };
-            renderer.line(MessageStyle::Info, message)?;
-        }
-        ActivePalette::Sessions { .. } => {
-            renderer.line(MessageStyle::Info, "Closed session browser.")?;
-        }
-        ActivePalette::Help => {
-            renderer.line(MessageStyle::Info, "Dismissed slash command help.")?;
-        }
-    }
-    Ok(())
-}
-
-struct CuratedPromptSection {
-    heading: &'static str,
-    component: ContextComponent,
-    body: String,
-}
-
-fn map_role_to_component(role: &uni::MessageRole) -> ContextComponent {
-    match role {
-        uni::MessageRole::System => ContextComponent::SystemPrompt,
-        uni::MessageRole::User => ContextComponent::UserMessage,
-        uni::MessageRole::Assistant => ContextComponent::AssistantMessage,
-        uni::MessageRole::Tool => ContextComponent::ToolResult,
-    }
-}
-
-fn describe_phase(phase: ConversationPhase) -> Option<String> {
-    match phase {
-        ConversationPhase::Exploration => Some("Exploration – gathering context".to_string()),
-        ConversationPhase::Implementation => {
-            Some("Implementation – applying code changes".to_string())
-        }
-        ConversationPhase::Validation => {
-            Some("Validation – executing tests and checks".to_string())
-        }
-        ConversationPhase::Debugging => {
-            Some("Debugging – addressing failures or regressions".to_string())
-        }
-        ConversationPhase::Unknown => None,
-    }
-}
-
-fn resolve_mode_label(preference: UiSurfacePreference, full_auto: bool) -> String {
-    let base = match preference {
-        UiSurfacePreference::Alternate => ui::HEADER_MODE_ALTERNATE,
-        UiSurfacePreference::Inline => ui::HEADER_MODE_INLINE,
-        UiSurfacePreference::Auto => ui::HEADER_MODE_AUTO,
-    };
-    if full_auto {
-        format!("{}{}", base, ui::HEADER_MODE_FULL_AUTO_SUFFIX)
-    } else {
-        base.to_string()
-    }
-}
-
-fn format_provider_label(value: &str) -> String {
-    let trimmed = value.trim();
-    if trimmed.is_empty() {
-        return String::new();
-    }
-
-    Provider::from_str(trimmed)
-        .map(|provider| provider.label().to_string())
-        .unwrap_or_else(|_| {
-            let mut chars = trimmed.chars();
-            let Some(first) = chars.next() else {
-                return String::new();
-            };
-            let mut formatted: String = first.to_uppercase().collect();
-            formatted.push_str(chars.as_str());
-            formatted
-        })
-}
-
-fn build_curated_sections(context: &CuratedContext) -> Vec<CuratedPromptSection> {
-    let mut sections = Vec::new();
-
-    if let Some(summary) = &context.ledger_summary {
-        if !summary.trim().is_empty() {
-            sections.push(CuratedPromptSection {
-                heading: "Decision Ledger",
-                component: ContextComponent::DecisionLedger,
-                body: summary.trim().to_string(),
-            });
-        }
-    }
-
-    if !context.active_files.is_empty() {
-        let mut body = String::new();
-        for file in &context.active_files {
-            let _ = writeln!(body, "{} ({} lines)", file.path, file.size_lines);
-            if !file.summary.trim().is_empty() {
-                let _ = writeln!(body, "  {}", file.summary.trim());
-            }
-        }
-        sections.push(CuratedPromptSection {
-            heading: "Active Files",
-            component: ContextComponent::FileContent,
-            body,
-        });
-    }
-
-    if !context.recent_errors.is_empty() {
-        let mut body = String::new();
-        for error in &context.recent_errors {
-            let mut line = error.error_message.trim().to_string();
-            if let Some(tool) = &error.tool_name {
-                line.push_str(&format!(" (tool: {})", tool));
-            }
-            let _ = writeln!(body, "{}", line);
-            if let Some(resolution) = &error.resolution {
-                if !resolution.trim().is_empty() {
-                    let _ = writeln!(body, "  resolution: {}", resolution.trim());
-                }
-            }
-        }
-        sections.push(CuratedPromptSection {
-            heading: "Recent Errors",
-            component: ContextComponent::ToolResult,
-            body,
-        });
-    }
-
-    if !context.relevant_tools.is_empty() {
-        let mut body = String::new();
-        for tool in &context.relevant_tools {
-            let description = tool.description.trim();
-            if description.is_empty() {
-                let _ = writeln!(body, "{}", tool.name);
-            } else {
-                let _ = writeln!(body, "{} – {}", tool.name, description);
-            }
-        }
-        sections.push(CuratedPromptSection {
-            heading: "Relevant Tools",
-            component: ContextComponent::ProjectGuidelines,
-            body,
-        });
-    }
-
-    if let Some(phase_text) = describe_phase(context.phase) {
-        sections.push(CuratedPromptSection {
-            heading: "Conversation Phase",
-            component: ContextComponent::ProjectGuidelines,
-            body: phase_text,
-        });
-    }
-
-    sections
-}
-
-async fn build_curator_messages(
-    history: &[uni::Message],
-    token_budget: &TokenBudgetManager,
-    token_budget_enabled: bool,
-) -> Result<Vec<CuratorMessage>> {
-    let mut messages = Vec::with_capacity(history.len());
-
-    for (index, message) in history.iter().enumerate() {
-        let mut materialized = message.content.clone();
-        if let Some(tool_calls) = &message.tool_calls {
-            if !tool_calls.is_empty() {
-                let serialized =
-                    serde_json::to_string(tool_calls).unwrap_or_else(|_| "[]".to_string());
-                if !serialized.is_empty() {
-                    if !materialized.is_empty() {
-                        materialized.push('\n');
-                    }
-                    materialized.push_str(&serialized);
-                }
-            }
-        }
-
-        let component = map_role_to_component(&message.role);
-        let component_id = format!("msg_{}", index);
-        let component_id_ref = Some(component_id.as_str());
-        let estimated_tokens = if token_budget_enabled {
-            match token_budget
-                .count_tokens_for_component(&materialized, component, component_id_ref)
-                .await
-            {
-                Ok(count) => count,
-                Err(err) => {
-                    warn!(
-                        ?err,
-                        "Failed to count tokens for conversation message; using rough estimate"
-                    );
-                    let estimate = materialized.len() / 4;
-                    token_budget
-                        .record_tokens_for_component(component, estimate, component_id_ref)
-                        .await;
-                    estimate
-                }
-            }
-        } else {
-            materialized.len() / 4
-        };
-
-        messages.push(CuratorMessage {
-            role: message.role.as_generic_str().to_string(),
-            content: materialized,
-            estimated_tokens,
-        });
-    }
-
-    Ok(messages)
-}
-
-fn build_curator_tools(tools: &[uni::ToolDefinition]) -> Vec<CuratorToolDefinition> {
-    tools
-        .iter()
-        .map(|tool| {
-            let parameters_repr = tool.function.parameters.to_string();
-            let estimated_tokens = tool.function.description.len() / 4 + parameters_repr.len() / 4;
-            CuratorToolDefinition {
-                name: tool.function.name.clone(),
-                description: tool.function.description.clone(),
-                estimated_tokens,
-            }
-        })
-        .collect()
-}
-
-fn finalize_model_selection(
-    renderer: &mut AnsiRenderer,
-    picker: &ModelPickerState,
-    selection: ModelSelectionResult,
-    config: &mut CoreAgentConfig,
-    vt_cfg: &mut Option<VTCodeConfig>,
-    provider_client: &mut Box<dyn uni::LLMProvider>,
-    session_bootstrap: &SessionBootstrap,
-    handle: &InlineHandle,
-    full_auto: bool,
-) -> Result<()> {
-    let workspace = config.workspace.clone();
-
-    let api_key = if let Some(key) = selection.api_key.as_ref() {
-        persist_env_value(&workspace, &selection.env_key, key)?;
-        unsafe {
-            // SAFETY: we only write ASCII-alphanumeric keys derived from known providers or
-            // sanitized user input, and values are supplied directly by the user.
-            std::env::set_var(&selection.env_key, key);
-        }
-        key.clone()
-    } else if selection.provider_enum.is_some() {
-        let key = get_api_key(&selection.provider, &ApiKeySources::default())
-            .with_context(|| format!("API key not found for provider '{}'", selection.provider))?;
-        unsafe {
-            // SAFETY: see above. Keys are sanitized and values come from configuration sources.
-            std::env::set_var(&selection.env_key, &key);
-        }
-        key
-    } else {
-        match std::env::var(&selection.env_key) {
-            Ok(value) if !value.trim().is_empty() => value,
-            _ if selection.requires_api_key => {
-                return Err(anyhow!(
-                    "API key not found for provider '{}'. Set {} or enter a key to continue.",
-                    selection.provider,
-                    selection.env_key
-                ));
-            }
-            _ => String::new(),
-        }
-    };
-
-    if let Some(provider_enum) = selection.provider_enum {
-        if let Err(err) = verify_model_with_rig(provider_enum, &selection.model, &api_key) {
-            renderer.line(
-                MessageStyle::Error,
-                &format!(
-                    "Rig validation warning: unable to initialise {} via rig-core ({err}).",
-                    selection.model_display
-                ),
-            )?;
-        }
-    }
-
-    let updated_cfg = picker.persist_selection(&workspace, &selection)?;
-    *vt_cfg = Some(updated_cfg);
-
-    if let Some(provider_enum) = selection.provider_enum {
-        let provider_name = selection.provider.clone();
-        let new_client = create_provider_with_config(
-            &provider_name,
-            Some(api_key.clone()),
-            None,
-            Some(selection.model.clone()),
-            Some(config.prompt_cache.clone()),
-        )
-        .context("Failed to initialize provider for the selected model")?;
-        *provider_client = new_client;
-        config.provider = provider_enum.to_string();
-    } else {
-        renderer.line(
-            MessageStyle::Info,
-            "Saved selection, but custom providers require manual configuration before taking effect.",
-        )?;
-        config.provider = selection.provider.clone();
-    }
-
-    config.model = selection.model.clone();
-    config.api_key = api_key;
-    config.reasoning_effort = selection.reasoning;
-    config.api_key_env = selection.env_key.clone();
-    if let Some(ref key) = selection.api_key {
-        config
-            .custom_api_keys
-            .insert(selection.provider.clone(), key.clone());
-    } else {
-        config.custom_api_keys.remove(&selection.provider);
-    }
-
-    if let Some(provider_enum) = selection.provider_enum {
-        if selection.reasoning_supported {
-            if let Some(payload) = reasoning_parameters_for(provider_enum, selection.reasoning) {
-                renderer.line(
-                    MessageStyle::Info,
-                    &format!("Rig reasoning configuration prepared: {}", payload),
-                )?;
-            }
-        }
-    }
-
-    let reasoning_label = selection.reasoning.as_str().to_string();
-    let mode_label = resolve_mode_label(config.ui_surface, full_auto);
-    let header_context = build_inline_header_context(
-        config,
-        session_bootstrap,
-        selection.provider_label.clone(),
-        selection.model.clone(),
-        mode_label,
-        reasoning_label.clone(),
-    )?;
-    handle.set_header_context(header_context);
-
-    renderer.line(
-        MessageStyle::Info,
-        &format!(
-            "Model set to {} ({}) via {}.",
-            selection.model_display, selection.model, selection.provider_label
-        ),
-    )?;
-
-    if !selection.known_model {
-        renderer.line(
-            MessageStyle::Info,
-            "The selected model is not part of VTCode's curated list; capabilities may vary.",
-        )?;
-    }
-
-    if selection.reasoning_supported {
-        let message = if selection.reasoning_changed {
-            format!("Reasoning effort updated to '{}'.", selection.reasoning)
-        } else {
-            format!("Reasoning effort remains '{}'.", selection.reasoning)
-        };
-        renderer.line(MessageStyle::Info, &message)?;
-    }
-
-    if selection.api_key.is_some() {
-        renderer.line(
-            MessageStyle::Info,
-            &format!(
-                "Stored credential under {} and updated the active environment.",
-                selection.env_key
-            ),
-        )?;
-    } else if selection.requires_api_key {
-        renderer.line(
-            MessageStyle::Info,
-            &format!(
-                "Using environment variable {} for authentication.",
-                selection.env_key
-            ),
-        )?;
-    }
-
-    Ok(())
-}
-
-fn persist_env_value(workspace: &Path, key: &str, value: &str) -> Result<()> {
-    let env_path = workspace.join(".env");
-    let mut lines: Vec<String> = if env_path.exists() {
-        std::fs::read_to_string(&env_path)
-            .with_context(|| format!("Failed to read {}", env_path.display()))?
-            .lines()
-            .map(|line| line.to_string())
-            .collect()
-    } else {
-        Vec::new()
-    };
-
-    let mut replaced = false;
-    for line in lines.iter_mut() {
-        if let Some((existing_key, _)) = line.split_once('=') {
-            if existing_key.trim() == key {
-                *line = format!("{key}={value}");
-                replaced = true;
-            }
-        }
-    }
-
-    if !replaced {
-        lines.push(format!("{key}={value}"));
-    }
-
-    let parent = env_path
-        .parent()
-        .map(std::path::Path::to_path_buf)
-        .unwrap_or_else(|| workspace.to_path_buf());
-
-    if !parent.exists() {
-        std::fs::create_dir_all(&parent)
-            .with_context(|| format!("Failed to create directory {}", parent.display()))?;
-    }
-
-    let temp = Builder::new()
-        .prefix(".env.")
-        .suffix(".tmp")
-        .tempfile_in(&parent)
-        .with_context(|| format!("Failed to create temporary file in {}", parent.display()))?;
-
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-        let permissions = std::fs::Permissions::from_mode(0o600);
-        temp.as_file()
-            .set_permissions(permissions)
-            .with_context(|| format!("Failed to set permissions on {}", temp.path().display()))?;
-    }
-
-    {
-        let mut writer = BufWriter::new(temp.as_file());
-        for line in &lines {
-            writeln!(writer, "{line}")
-                .with_context(|| format!("Failed to write .env entry for {key}"))?;
-        }
-        writer
-            .flush()
-            .with_context(|| format!("Failed to flush temporary .env for {}", key))?;
-    }
-
-    temp.as_file()
-        .sync_all()
-        .with_context(|| format!("Failed to sync temporary .env for {}", key))?;
-
-    let _file = temp
-        .persist(&env_path)
-        .with_context(|| format!("Failed to persist {}", env_path.display()))?;
-
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-        std::fs::set_permissions(&env_path, std::fs::Permissions::from_mode(0o600))
-            .with_context(|| format!("Failed to set permissions on {}", env_path.display()))?;
-    }
-
-    Ok(())
-}
-
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 enum HitlDecision {
     Approved,
@@ -1905,580 +273,10 @@ impl PlaceholderGuard {
     }
 }
 
-#[derive(Default, Clone)]
-struct InputStatusState {
-    left: Option<String>,
-    right: Option<String>,
-    git_left: Option<String>,
-    git_summary: Option<GitStatusSummary>,
-    last_git_refresh: Option<Instant>,
-    command_value: Option<String>,
-    last_command_refresh: Option<Instant>,
-}
-
-const GIT_STATUS_REFRESH_INTERVAL: Duration = Duration::from_secs(2);
-
-async fn update_input_status_if_changed(
-    handle: &InlineHandle,
-    workspace: &Path,
-    model: &str,
-    reasoning: &str,
-    status_config: Option<&StatusLineConfig>,
-    state: &mut InputStatusState,
-) -> Result<()> {
-    let mode = status_config
-        .map(|cfg| cfg.mode)
-        .unwrap_or(StatusLineMode::Auto);
-
-    if matches!(mode, StatusLineMode::Hidden) {
-        state.last_git_refresh = None;
-        state.git_left = None;
-        state.git_summary = None;
-    } else {
-        let should_refresh_git = match state.last_git_refresh {
-            Some(last_refresh) => last_refresh.elapsed() >= GIT_STATUS_REFRESH_INTERVAL,
-            None => true,
-        };
-
-        if should_refresh_git {
-            match git_status_summary(workspace) {
-                Ok(Some(summary)) => {
-                    let mut branch = summary.branch.clone();
-                    if summary.dirty {
-                        branch.push('*');
-                    }
-                    state.git_left = Some(branch);
-                    state.git_summary = Some(summary);
-                }
-                Ok(None) => {
-                    state.git_left = None;
-                    state.git_summary = None;
-                }
-                Err(error) => {
-                    warn!(
-                        workspace = %workspace.display(),
-                        error = ?error,
-                        "Failed to resolve git status"
-                    );
-                    state.git_summary = None;
-                }
-            }
-
-            state.last_git_refresh = Some(Instant::now());
-        }
-    }
-
-    let trimmed_model = model.trim();
-    let trimmed_reasoning = reasoning.trim();
-    let model_display = ModelId::from_str(trimmed_model)
-        .map(|id| id.display_name().to_string())
-        .unwrap_or_else(|_| trimmed_model.to_string());
-
-    let mut command_error: Option<anyhow::Error> = None;
-
-    let (left, right) = match mode {
-        StatusLineMode::Hidden => {
-            state.command_value = None;
-            state.last_command_refresh = None;
-            (None, None)
-        }
-        StatusLineMode::Auto => {
-            let right = build_model_status_right(trimmed_model, trimmed_reasoning);
-            (state.git_left.clone(), right)
-        }
-        StatusLineMode::Command => {
-            if let Some(cfg) = status_config {
-                if let Some(command) = cfg
-                    .command
-                    .as_ref()
-                    .map(|cmd| cmd.trim().to_string())
-                    .filter(|cmd| !cmd.is_empty())
-                {
-                    let refresh_interval = Duration::from_millis(cfg.refresh_interval_ms);
-                    let should_refresh_command = match state.last_command_refresh {
-                        Some(last_refresh) => {
-                            if refresh_interval.is_zero() {
-                                true
-                            } else {
-                                last_refresh.elapsed() >= refresh_interval
-                            }
-                        }
-                        None => true,
-                    };
-
-                    if should_refresh_command {
-                        state.last_command_refresh = Some(Instant::now());
-                        match run_status_line_command(
-                            &command,
-                            workspace,
-                            trimmed_model,
-                            &model_display,
-                            trimmed_reasoning,
-                            state.git_summary.as_ref(),
-                            cfg,
-                        )
-                        .await
-                        {
-                            Ok(output) => {
-                                state.command_value = output;
-                            }
-                            Err(error) => {
-                                command_error = Some(error);
-                            }
-                        }
-                    }
-
-                    (state.command_value.clone(), None)
-                } else {
-                    state.command_value = None;
-                    let right = build_model_status_right(trimmed_model, trimmed_reasoning);
-                    (state.git_left.clone(), right)
-                }
-            } else {
-                state.command_value = None;
-                let right = build_model_status_right(trimmed_model, trimmed_reasoning);
-                (state.git_left.clone(), right)
-            }
-        }
-    };
-
-    if state.left != left || state.right != right {
-        handle.set_input_status(left.clone(), right.clone());
-        state.left = left;
-        state.right = right;
-    }
-
-    if let Some(error) = command_error {
-        Err(error)
-    } else {
-        Ok(())
-    }
-}
-
-fn build_model_status_right(model: &str, reasoning: &str) -> Option<String> {
-    if model.is_empty() {
-        None
-    } else if reasoning.is_empty() {
-        Some(model.to_string())
-    } else {
-        Some(format!("{} ({})", model, reasoning))
-    }
-}
-
-async fn run_status_line_command(
-    command: &str,
-    workspace: &Path,
-    model_id: &str,
-    model_display: &str,
-    reasoning: &str,
-    git: Option<&GitStatusSummary>,
-    config: &StatusLineConfig,
-) -> Result<Option<String>> {
-    let mut process = TokioCommand::new("sh");
-    process.arg("-c").arg(command);
-    process.current_dir(workspace);
-    process.stdin(std::process::Stdio::piped());
-    process.stdout(std::process::Stdio::piped());
-    process.stderr(std::process::Stdio::null());
-
-    let mut child = process
-        .spawn()
-        .with_context(|| format!("failed to spawn status line command `{}`", command))?;
-
-    let mut stdout_pipe = child
-        .stdout
-        .take()
-        .context("status line command missing stdout pipe")?;
-
-    if let Some(mut stdin) = child.stdin.take() {
-        let payload =
-            StatusLineCommandPayload::new(workspace, model_id, model_display, reasoning, git);
-        let mut payload_bytes =
-            serde_json::to_vec(&payload).context("failed to serialize status line payload")?;
-        payload_bytes.push(b'\n');
-
-        stdin
-            .write_all(&payload_bytes)
-            .await
-            .context("failed to write status line payload")?;
-        stdin
-            .shutdown()
-            .await
-            .context("failed to close status line command stdin")?;
-    }
-
-    let timeout_ms = std::cmp::max(config.command_timeout_ms, 1);
-    let timeout_duration = Duration::from_millis(timeout_ms);
-    let wait_result = {
-        let mut wait = child.wait();
-        tokio::pin!(wait);
-        tokio::time::timeout(timeout_duration, &mut wait).await
-    };
-
-    let status = match wait_result {
-        Ok(status_res) => status_res
-            .with_context(|| format!("failed to wait for status line command `{}`", command))?,
-        Err(_) => {
-            child
-                .start_kill()
-                .with_context(|| format!("failed to kill timed out status line command `{}`", command))?;
-            child.wait().await.with_context(|| {
-                format!(
-                    "failed to wait for killed status line command `{}` after timeout",
-                    command
-                )
-            })?;
-            return Err(anyhow!(
-                "status line command `{}` timed out after {}ms",
-                command,
-                timeout_ms
-            ));
-        }
-    };
-
-    let mut stdout_bytes = Vec::new();
-    stdout_pipe
-        .read_to_end(&mut stdout_bytes)
-        .await
-        .context("failed to read status line command stdout")?;
-
-    if !status.success() {
-        return Err(anyhow!("status line command exited with status {}", status));
-    }
-
-    let stdout = String::from_utf8_lossy(&stdout_bytes);
-    let first_line = stdout
-        .lines()
-        .next()
-        .map(|line| line.trim_end().to_string())
-        .filter(|line| !line.is_empty());
-
-    Ok(first_line)
-}
-
-#[derive(Serialize)]
-struct StatusLineCommandPayload {
-    hook_event_name: &'static str,
-    cwd: String,
-    workspace: StatusLineWorkspace,
-    model: StatusLineModel,
-    runtime: StatusLineRuntime,
-    git: Option<StatusLineGit>,
-    version: &'static str,
-}
-
-impl StatusLineCommandPayload {
-    fn new(
-        workspace: &Path,
-        model_id: &str,
-        model_display: &str,
-        reasoning: &str,
-        git: Option<&GitStatusSummary>,
-    ) -> Self {
-        let workspace_path = workspace.to_string_lossy().to_string();
-        Self {
-            hook_event_name: "Status",
-            cwd: workspace_path.clone(),
-            workspace: StatusLineWorkspace {
-                current_dir: workspace_path.clone(),
-                project_dir: workspace_path.clone(),
-            },
-            model: StatusLineModel {
-                id: model_id.to_string(),
-                display_name: model_display.to_string(),
-            },
-            runtime: StatusLineRuntime {
-                reasoning_effort: reasoning.to_string(),
-            },
-            git: git.map(StatusLineGit::from_summary),
-            version: env!("CARGO_PKG_VERSION"),
-        }
-    }
-}
-
-#[derive(Serialize)]
-struct StatusLineWorkspace {
-    current_dir: String,
-    project_dir: String,
-}
-
-#[derive(Serialize)]
-struct StatusLineModel {
-    id: String,
-    display_name: String,
-}
-
-#[derive(Serialize)]
-struct StatusLineRuntime {
-    reasoning_effort: String,
-}
-
-#[derive(Serialize)]
-struct StatusLineGit {
-    branch: String,
-    dirty: bool,
-}
-
-impl StatusLineGit {
-    fn from_summary(summary: &GitStatusSummary) -> Self {
-        Self {
-            branch: summary.branch.clone(),
-            dirty: summary.dirty,
-        }
-    }
-}
-
 impl Drop for PlaceholderGuard {
     fn drop(&mut self) {
         self.handle.set_placeholder(self.restore.clone());
     }
-}
-
-fn render_tool_call_summary(
-    renderer: &mut AnsiRenderer,
-    tool_name: &str,
-    args: &Value,
-) -> Result<()> {
-    let (headline, _) = describe_tool_action(tool_name, args);
-    renderer.line(MessageStyle::Info, &format!("→ {}", headline))?;
-
-    Ok(())
-}
-
-fn describe_tool_action(tool_name: &str, args: &Value) -> (String, HashSet<String>) {
-    match tool_name {
-        tool_names::RUN_TERMINAL_CMD | tool_names::BASH => describe_shell_command(args)
-            .unwrap_or_else(|| ("Run shell command".to_string(), HashSet::new())),
-        tool_names::LIST_FILES => {
-            describe_list_files(args).unwrap_or_else(|| ("List files".to_string(), HashSet::new()))
-        }
-        tool_names::GREP_SEARCH => describe_grep_search(args)
-            .unwrap_or_else(|| ("Search with grep".to_string(), HashSet::new())),
-        tool_names::READ_FILE => describe_path_action(args, "Read file", &["path"])
-            .unwrap_or_else(|| ("Read file".to_string(), HashSet::new())),
-        tool_names::WRITE_FILE => describe_path_action(args, "Write file", &["path"])
-            .unwrap_or_else(|| ("Write file".to_string(), HashSet::new())),
-        tool_names::EDIT_FILE => describe_path_action(args, "Edit file", &["path"])
-            .unwrap_or_else(|| ("Edit file".to_string(), HashSet::new())),
-        tool_names::CREATE_FILE => describe_path_action(args, "Create file", &["path"])
-            .unwrap_or_else(|| ("Create file".to_string(), HashSet::new())),
-        tool_names::DELETE_FILE => describe_path_action(args, "Delete file", &["path"])
-            .unwrap_or_else(|| ("Delete file".to_string(), HashSet::new())),
-        tool_names::CURL => {
-            describe_curl(args).unwrap_or_else(|| ("Fetch URL".to_string(), HashSet::new()))
-        }
-        tool_names::SIMPLE_SEARCH => describe_simple_search(args)
-            .unwrap_or_else(|| ("Search workspace".to_string(), HashSet::new())),
-        tool_names::SRGN => describe_srgn(args)
-            .unwrap_or_else(|| ("Search and replace".to_string(), HashSet::new())),
-        tool_names::APPLY_PATCH => ("Apply workspace patch".to_string(), HashSet::new()),
-        tool_names::UPDATE_PLAN => ("Update task plan".to_string(), HashSet::new()),
-        _ => (
-            format!("Use {}", humanize_tool_name(tool_name)),
-            HashSet::new(),
-        ),
-    }
-}
-
-fn describe_shell_command(args: &Value) -> Option<(String, HashSet<String>)> {
-    let mut used = HashSet::new();
-    if let Some(parts) = args
-        .get("command")
-        .and_then(|value| value.as_array())
-        .map(|array| {
-            array
-                .iter()
-                .filter_map(|value| value.as_str().map(str::to_string))
-                .collect::<Vec<_>>()
-        })
-        .filter(|parts: &Vec<String>| !parts.is_empty())
-    {
-        used.insert("command".to_string());
-        let joined = parts.join(" ");
-        let summary = truncate_middle(&joined, 60);
-        return Some((format!("Run command {}", summary), used));
-    }
-
-    if let Some(cmd) = args
-        .get("bash_command")
-        .and_then(|value| value.as_str())
-        .filter(|s| !s.is_empty())
-    {
-        used.insert("bash_command".to_string());
-        let summary = truncate_middle(cmd, 60);
-        return Some((format!("Run bash {}", summary), used));
-    }
-
-    None
-}
-
-fn describe_list_files(args: &Value) -> Option<(String, HashSet<String>)> {
-    if let Some(path) = lookup_string(args, "path") {
-        let mut used = HashSet::new();
-        used.insert("path".to_string());
-        let location = if path == "." {
-            "workspace root".to_string()
-        } else {
-            truncate_middle(&path, 60)
-        };
-        return Some((format!("List files in {}", location), used));
-    }
-    if let Some(pattern) = lookup_string(args, "name_pattern") {
-        let mut used = HashSet::new();
-        used.insert("name_pattern".to_string());
-        return Some((
-            format!("Find files named {}", truncate_middle(&pattern, 40)),
-            used,
-        ));
-    }
-    if let Some(pattern) = lookup_string(args, "content_pattern") {
-        let mut used = HashSet::new();
-        used.insert("content_pattern".to_string());
-        return Some((
-            format!("Search files for {}", truncate_middle(&pattern, 40)),
-            used,
-        ));
-    }
-    None
-}
-
-fn describe_grep_search(args: &Value) -> Option<(String, HashSet<String>)> {
-    let pattern = lookup_string(args, "pattern");
-    let path = lookup_string(args, "path");
-    match (pattern, path) {
-        (Some(pat), Some(path)) => {
-            let mut used = HashSet::new();
-            used.insert("pattern".to_string());
-            used.insert("path".to_string());
-            Some((
-                format!(
-                    "Grep {} in {}",
-                    truncate_middle(&pat, 40),
-                    truncate_middle(&path, 40)
-                ),
-                used,
-            ))
-        }
-        (Some(pat), None) => {
-            let mut used = HashSet::new();
-            used.insert("pattern".to_string());
-            Some((format!("Grep {}", truncate_middle(&pat, 40)), used))
-        }
-        _ => None,
-    }
-}
-
-fn describe_simple_search(args: &Value) -> Option<(String, HashSet<String>)> {
-    if let Some(query) = lookup_string(args, "query") {
-        let mut used = HashSet::new();
-        used.insert("query".to_string());
-        return Some((format!("Search for {}", truncate_middle(&query, 50)), used));
-    }
-    None
-}
-
-fn describe_srgn(args: &Value) -> Option<(String, HashSet<String>)> {
-    let pattern = lookup_string(args, "pattern");
-    let replacement = lookup_string(args, "replacement");
-    match (pattern, replacement) {
-        (Some(pat), Some(rep)) => {
-            let mut used = HashSet::new();
-            used.insert("pattern".to_string());
-            used.insert("replacement".to_string());
-            Some((
-                format!(
-                    "Replace {} → {}",
-                    truncate_middle(&pat, 30),
-                    truncate_middle(&rep, 30)
-                ),
-                used,
-            ))
-        }
-        (Some(pat), None) => {
-            let mut used = HashSet::new();
-            used.insert("pattern".to_string());
-            Some((format!("Search for {}", truncate_middle(&pat, 40)), used))
-        }
-        _ => None,
-    }
-}
-
-fn describe_path_action(
-    args: &Value,
-    verb: &str,
-    keys: &[&str],
-) -> Option<(String, HashSet<String>)> {
-    for key in keys {
-        if let Some(value) = lookup_string(args, key) {
-            let mut used = HashSet::new();
-            used.insert((*key).to_string());
-            let summary = truncate_middle(&value, 60);
-            return Some((format!("{} {}", verb, summary), used));
-        }
-    }
-    None
-}
-
-fn describe_curl(args: &Value) -> Option<(String, HashSet<String>)> {
-    if let Some(url) = lookup_string(args, "url") {
-        let mut used = HashSet::new();
-        used.insert("url".to_string());
-        return Some((format!("Fetch {}", truncate_middle(&url, 60)), used));
-    }
-    None
-}
-
-fn lookup_string(args: &Value, key: &str) -> Option<String> {
-    args.as_object()
-        .and_then(|map| map.get(key))
-        .and_then(|value| value.as_str())
-        .map(|s| s.to_string())
-        .filter(|s| !s.is_empty())
-}
-fn humanize_tool_name(name: &str) -> String {
-    humanize_key(name)
-}
-
-fn humanize_key(key: &str) -> String {
-    let replaced = key.replace('_', " ");
-    if replaced.is_empty() {
-        return replaced;
-    }
-    let mut chars = replaced.chars();
-    let first = chars.next().unwrap();
-    let mut result = first.to_uppercase().collect::<String>();
-    result.push_str(&chars.collect::<String>());
-    result
-}
-
-fn truncate_middle(text: &str, max_len: usize) -> String {
-    if max_len == 0 {
-        return String::new();
-    }
-    let chars: Vec<char> = text.chars().collect();
-    if chars.len() <= max_len {
-        return text.to_string();
-    }
-    if max_len <= 1 {
-        return "…".to_string();
-    }
-    let head_len = max_len / 2;
-    let tail_len = max_len.saturating_sub(head_len + 1);
-    let mut result: String = chars.iter().take(head_len).collect();
-    result.push('…');
-    if tail_len > 0 {
-        let tail: String = chars
-            .iter()
-            .rev()
-            .take(tail_len)
-            .cloned()
-            .collect::<Vec<char>>()
-            .into_iter()
-            .rev()
-            .collect();
-        result.push_str(&tail);
-    }
-    result
 }
 
 async fn prompt_tool_permission(
@@ -2582,7 +380,7 @@ async fn prompt_tool_permission(
 }
 
 async fn ensure_tool_permission(
-    tool_registry: &mut vtcode_core::tools::registry::ToolRegistry,
+    tool_registry: &mut ToolRegistry,
     tool_name: &str,
     renderer: &mut AnsiRenderer,
     handle: &InlineHandle,
@@ -2646,12 +444,6 @@ async fn ensure_tool_permission(
             }
         }
     }
-}
-
-fn apply_prompt_style(handle: &InlineHandle) {
-    let styles = theme::active_styles();
-    let style = convert_ui_style(styles.primary);
-    handle.set_prompt("❯ ".to_string(), style);
 }
 
 const SPINNER_UPDATE_INTERVAL_MS: u64 = 120;
@@ -3052,10 +844,10 @@ async fn stream_and_render_response(
             reasoning_state
                 .handle_stream_failure(renderer)
                 .map_err(|err| map_render_error(provider_name, err))?;
-        let formatted_error = error_display::format_llm_error(
-            provider_name,
-            "Stream ended without a completion event",
-        );
+            let formatted_error = error_display::format_llm_error(
+                provider_name,
+                "Stream ended without a completion event",
+            );
             return Err(uni::LLMError::Provider(formatted_error));
         }
     };
